move utils.new_nii_like to io.new_nii_like
ref: https://github.com/ME-ICA/tedana/issues/135
