+++ tedana/io.py
@@ -5,7 +5,10 @@ import logging
 import os.path as op
 import textwrap
 
+import nibabel as nib
 import numpy as np
+from nilearn._utils import check_niimg
+from nilearn.image import new_img_like
 from numpy.linalg import lstsq
 
 from tedana import model, utils
@@ -513,3 +516,39 @@ def ctabsel(ctabfile):
             if ll[:4] is kk and ll[4:].strip() is not '':
                 class_dict[kk] = ll[4:].split('#')[0].split(',')
     return tuple([np.array(class_dict[kk], dtype=int) for kk in class_tags])
+
+
+def new_nii_like(ref_img, data, affine=None, copy_header=True):
+    """
+    Coerces `data` into NiftiImage format like `ref_img`
+
+    Parameters
+    ----------
+    ref_img : :obj:`str` or img_like
+        Reference image
+    data : (S [x T]) array_like
+        Data to be saved
+    affine : (4 x 4) array_like, optional
+        Transformation matrix to be used. Default: `ref_img.affine`
+    copy_header : :obj:`bool`, optional
+        Whether to copy header from `ref_img` to new image. Default: True
+
+    Returns
+    -------
+    nii : :obj:`nibabel.nifti1.Nifti1Image`
+        NiftiImage
+    """
+
+    ref_img = check_niimg(ref_img)
+    newdata = data.reshape(ref_img.shape[:3] + data.shape[1:])
+    if '.nii' not in ref_img.valid_exts:
+        # this is rather ugly and may lose some information...
+        nii = nib.Nifti1Image(newdata, affine=ref_img.affine,
+                              header=ref_img.header)
+    else:
+        # nilearn's `new_img_like` is a very nice function
+        nii = new_img_like(ref_img, newdata, affine=affine,
+                           copy_header=copy_header)
+    nii.set_data_dtype(data.dtype)
+
+    return nii
